+++ R/fst_WC84.R
@@ -44,12 +44,12 @@
 
 #' @param data A file in the working directory or object in the global environment 
 #' in wide or long (tidy) formats. To import, the function uses internally
-#' \href{https://github.com/thierrygosselin/stackr}{stackr} 
-#' \code{\link[stackr]{tidy_wide}}. See details for more info.
+#' \href{https://github.com/thierrygosselin/radiator}{radiator} 
+#' \code{\link[radiator]{tidy_wide}}. See details for more info.
 #' 
 #' \emph{How to get a tidy data frame ?}
-#' \href{https://github.com/thierrygosselin/stackr}{stackr} 
-#' \code{\link[stackr]{tidy_genomic_data}} can transform 11 genomic data formats 
+#' \href{https://github.com/thierrygosselin/radiator}{radiator} 
+#' \code{\link[radiator]{tidy_genomic_data}} can transform 11 genomic data formats 
 #' in a tidy data frame (VCF, PLINK, genind, genlight, gtypes, genepop,
 #' stacks haplotype file, hierfstat, ...). 
 #' You can also use this function to filter your dataset using
@@ -159,7 +159,7 @@
 #' @details \strong{Input data:}
 #'  
 #' To discriminate the long from the wide format, 
-#' the function \pkg{stackr} \code{\link[stackr]{tidy_wide}} searches 
+#' the function \pkg{radiator} \code{\link[radiator]{tidy_wide}} searches 
 #' for \code{MARKERS or LOCUS} in column names (TRUE = long format).
 #' The data frame is tab delimitted.
 
@@ -171,7 +171,7 @@
 #' 
 #' \strong{Long/Tidy format:}
 #' The long format is considered to be a tidy data frame and can store metadata info. 
-#' (e.g. from a VCF see \pkg{stackr} \code{\link{tidy_genomic_data}}). A minimum of 4 columns
+#' (e.g. from a VCF see \pkg{radiator} \code{\link{tidy_genomic_data}}). A minimum of 4 columns
 #' are required in the long format: \code{INDIVIDUALS}, \code{POP_ID}, 
 #' \code{MARKERS or LOCUS} and \code{GENOTYPE or GT}. The rest are considered metata info.
 #' 
@@ -181,12 +181,12 @@
 #' The separator can be any of these: \code{"/", ":", "_", "-", "."}.
 #' 
 #' \emph{How to get a tidy data frame ?}
-#' \pkg{stackr} \code{\link{tidy_genomic_data}} can transform 6 genomic data formats 
+#' \pkg{radiator} \code{\link{tidy_genomic_data}} can transform 6 genomic data formats 
 #' in a tidy data frame.
 
 #' @export
 #' @rdname fst_WC84
-#' @importFrom stackr tidy_wide discard_monomorphic_markers keep_common_markers change_pop_names detect_biallelic_markers
+#' @importFrom radiator tidy_wide discard_monomorphic_markers keep_common_markers change_pop_names detect_biallelic_markers
 #' @importFrom tidyr separate gather spread unite
 #' @importFrom purrr map flatten transpose flatten_int
 #' @importFrom dplyr mutate summarise group_by ungroup select rename full_join left_join anti_join right_join semi_join filter n_distinct distinct arrange sample_n bind_rows bind_cols ntile desc n
@@ -268,7 +268,7 @@
 #' For Fisher's exact test and p-values per markers 
 #' see \code{mmod} \code{\link[mmod]{diff_test}}.
 #' 
-#' \code{\link[stackr]{tidy_genomic_data}} to transform numerous genomic data 
+#' \code{\link[radiator]{tidy_genomic_data}} to transform numerous genomic data 
 #' format in tidy data frames.
 
 #' @author Thierry Gosselin \email{thierrygosselin@@icloud.com}
@@ -327,7 +327,7 @@ fst_WC84 <- function(
   
   # Import data ---------------------------------------------------------------
   if (verbose) message("Importing data")
-  input <- stackr::tidy_wide(data = data, import.metadata = TRUE)
+  input <- radiator::tidy_wide(data = data, import.metadata = TRUE)
   
   # For long tidy format, switch LOCUS to MARKERS column name, if found MARKERS not found
   if (tibble::has_name(input, "LOCUS") && !tibble::has_name(input, "MARKERS")) {
@@ -369,7 +369,7 @@ fst_WC84 <- function(
   }
   
   # using pop.levels and pop.labels info if present
-  input <- stackr::change_pop_names(data = input, pop.levels = pop.levels, pop.labels = pop.labels)
+  input <- radiator::change_pop_names(data = input, pop.levels = pop.levels, pop.labels = pop.labels)
   
   # subsampling data------------------------------------------------------------
   # create the subsampling list
@@ -571,7 +571,7 @@ compute_fst <- function(x, ci = FALSE, iteration.ci = 100, quantiles.ci = c(0.02
   # x = data.genotyped # test
   
   # Removing monomorphic markers------------------------------------------------
-  x <- stackr::discard_monomorphic_markers(data = x, verbose = FALSE)$input
+  x <- radiator::discard_monomorphic_markers(data = x, verbose = FALSE)$input
   
   # number of marker used for computation 
   n.markers <- dplyr::n_distinct(x$MARKERS)
@@ -801,7 +801,7 @@ pairwise_fst <- function(
   
   pop.select <- stringi::stri_join(purrr::flatten(pop.pairwise[list.pair]))
   
-  # data.select <- stackr::keep_common_markers(data = data.select) # longer than below
+  # data.select <- radiator::keep_common_markers(data = data.select) # longer than below
   # common markers
   set1 <- unique.markers.pop %>%
     dplyr::filter(POP_ID == pop.select[1]) %>%
@@ -942,7 +942,7 @@ fst_subsample <- function(
   # SNPRelate and other prep ---------------------------------------------------
   # if (snprelate) {
   #   message("Detect if data is biallelic")
-  #   biallelic <- stackr::detect_biallelic_markers(data = input)
+  #   biallelic <- radiator::detect_biallelic_markers(data = input)
   #   
   #   if (!biallelic) {
   #     warning("Data is not biallelic = cannot run assigner with SNPRelate Fst function")
@@ -980,7 +980,7 @@ fst_subsample <- function(
   # if (snprelate) {
   #   if (verbose) message("Generating GDS format...")
   #   # keep markers in common
-  #   gds.genotypes <- suppressMessages(stackr::keep_common_markers(data = input))
+  #   gds.genotypes <- suppressMessages(radiator::keep_common_markers(data = input))
   #   
   #   strata.df <- dplyr::distinct(gds.genotypes, POP_ID, INDIVIDUALS) %>%
   #     dplyr::mutate(POP_ID = factor(POP_ID))